Search block signature matching at each byte offset.
Before, we were finding signature matches at block boundaries which
wasn't catching modifications at the top or middle of the file, causing
full file uploads. It now works but is slow. I'll be working on
making it work faster. Preeliminary profiling shows most of the time being
spent in the rollingHash function.